♻️ Refactor NFTPortfolioItem
