--- src/components/NFTPortfolio/Card.vue
@@ -1,23 +1,0 @@
-<template>
-  <div
-    :class="[
-      'flex',
-      'flex-col',
-      'rounded-[24px]',
-      'w-[310px]',
-      'mb-[16px]',
-      'overflow-hidden',
-      'w-full',
-      'bg-white',
-      'box-border',
-      'border-[2px]',
-      'border-transparent',
-      'hover:border-like-cyan',
-      'transition',
-      'ease-in',
-      'duration-200',
-    ]"
-  >
-    <slot />
-  </div>
-</template>
